requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
openai==1.54.4
anthropic==0.85.0
python-dateutil==2.8.2
pytz==2026.1.post1
flask==3.1.3
